Extract point plotting for UmapDemo to separate class PlotPoints
